kafka-fraud-detection/requirements.txt
kafka-python
sqlite3
